requirements.txt
streamlit==1.53.0
tensorflow-cpu==2.14.0
numpy
pandas
matplotlib
Pillow
